Refactor rainfall code out of init
